Add versioning and refactor entities
