+++ src/data/assets/modules/data-entry-module.json
@@ -1,4 +1,5 @@
 {
+    "_version": 1,
     "id": "data-entry",
     "disabled": false,
     "dhisAppKey": "data-entry",
