+++ src/data/clients/translation/TranslationClient.ts
@@ -1,11 +1,21 @@
 import { Either } from "../../../domain/entities/Either";
-import { TranslationProject } from "../../entities/Translation";
+
+export interface TranslationClient {
+    getProject(id: string): Promise<Either<TranslationError, TranslationProject>>;
+    createProject(): Promise<TranslationProject>;
+}
 
 export type TranslationProvider = "poeditor";
 
 export type TranslationError = "UNKNOWN";
 
-export interface TranslationClient {
-    getProject(id: string): Promise<Either<TranslationError, TranslationProject>>;
-    createProject(): Promise<TranslationProject>;
+export interface TranslationProject {
+    id: string;
+    name: string;
+    languages: TranslationLanguage[];
+}
+
+export interface TranslationLanguage {
+    id: string;
+    name: string;
 }
